Fix Audit Trail
This will be tested with the new task (#8)
